backend/requirements.txt
Django>=3.0,<4.0
djangorestframework
markdown
django-filter
psycopg2>=2.8
pandas
joblib
